Added eXvisory to bot building platforms
eXvisory is both a bot building platform (this PR) for building product support AI chatbots and a pilot chatbot for troubleshooting mobile phone and tablet problems. So PR #57 requests a listing as an example chatbot, while this PR requests listing as a bot building platform. If that is considered greedy I'd prefer the PR #57 listing please.